Move Grpc code to separate directory and split out code into more files
