+++ ApplicationCode/Application/CMakeLists_files.cmake
@@ -36,11 +36,6 @@ ${CMAKE_CURRENT_LIST_DIR}/RiaFontCache.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RiaEclipseFileNameTools.cpp
 )
 
-if (RESINSIGHT_ENABLE_GRPC)
-	list(APPEND SOURCE_GROUP_HEADER_FILES ${CMAKE_CURRENT_LIST_DIR}/RiaGrpcServer.h)
-	list(APPEND SOURCE_GROUP_SOURCE_FILES ${CMAKE_CURRENT_LIST_DIR}/RiaGrpcServer.cpp)
-endif(RESINSIGHT_ENABLE_GRPC)
-
 list(APPEND CODE_HEADER_FILES
 ${SOURCE_GROUP_HEADER_FILES}
 )
