+++ ApplicationCode/GrpcInterface/RiaGrpcServiceInterface.h
@@ -1,0 +1,35 @@
+/////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (C) 2019-     Equinor ASA
+//
+//  ResInsight is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  ResInsight is distributed in the hope that it will be useful, but WITHOUT ANY
+//  WARRANTY; without even the implied warranty of MERCHANTABILITY or
+//  FITNESS FOR A PARTICULAR PURPOSE.
+//
+//  See the GNU General Public License at <http://www.gnu.org/licenses/gpl.html>
+//  for more details.
+//
+//////////////////////////////////////////////////////////////////////////////////
+#pragma once
+
+#include <vector>
+
+class RiaGrpcServerCallMethod;
+
+namespace grpc
+{
+    class ServerCompletionQueue;    
+}
+
+class RiaGrpcServiceInterface
+{
+public:
+    virtual std::vector<RiaGrpcServerCallMethod*> createCallbacks(grpc::ServerCompletionQueue*) = 0;
+};
+
+
